Move profilePic to resolve circular import
