+++ src/pages/Profile.js
@@ -9,10 +9,11 @@ import {
   ShibbolethLink, spinnerOverlay, UnlinkFenceAccount
 } from 'src/components/common'
 import FooterWrapper from 'src/components/FooterWrapper'
-import { icon, profilePic, spinner } from 'src/components/icons'
+import { icon, spinner } from 'src/components/icons'
 import { TextInput, ValidatedInput } from 'src/components/input'
 import Modal from 'src/components/Modal'
 import { InfoBox } from 'src/components/PopupTrigger'
+import ProfilePicture from 'src/components/ProfilePicture'
 import { SimpleTabBar } from 'src/components/tabBars'
 import TopBar from 'src/components/TopBar'
 import { useWorkspaces } from 'src/components/workspace-utils'
@@ -531,7 +532,7 @@ const PersonalInfoTab = ({ setSaving }) => {
   return h(PageBox, { role: 'main', style: { flexGrow: 1 }, variant: PageBoxVariants.LIGHT }, [
     div({ style: styles.header.line }, [
       div({ style: { position: 'relative' } }, [
-        profilePic({ size: 48 }),
+        h(ProfilePicture, { size: 48 }),
         h(InfoBox, { style: { alignSelf: 'flex-end' } }, [
           'To change your profile image, visit your ',
           h(Link, {
